Move doesNotErrorWhenUsingApplicationMainClassProperty and checkLargeZipFilesWithZip64Enabled
